Implement IGNORE_INVISIBLE_LAYOUT Zest Style for Zest 2.0
With this change, invisible graph nodes and connections are ignored
within the LayoutContext and therefore within the layout algorithms.

Note that as part of this change, the Zest styles are no longer set
inside the Graph constructor, but via a separate setGraphStyle() method.
This is because the constructor styles are also used for creating the
underlying FigureCanvas, which would otherwise throw an exception due to
the unsupported styles.

While it technically breaks compatibility, it is practically impossible
to have used those styles up until now.

Fixes https://github.com/eclipse-gef/gef-classic/issues/608