Generalize PaletteView to accept generic IPageBookViewPage
This replaces the implicit cast to from a PalettePage to a
IPageBookViewPage and makes therefore restores the support for pages
that don't implement this marker interface.

Note: This is undefined behavior and custom pages that don't implement
this interface may be the cause of unforeseeable
problems in the future, given that this breaks the IAdaptable contract.

Resolves https://github.com/eclipse-gef/gef-classic/issues/649