Add support for mouse-wheel events to figures
This introduces a new MouseWheelListener interfaces that can be used by
figures to respond to mouse-wheel events.

In addition, this also solves the issue where the figure under the mouse
cursor was not updated, if it was moved via a scroll event, leading to
"mouseExited" and "mouseEntered" not being fired and other visual
artifacts such as tool-tips not disappearing.