Build GEF against both the SimRel version and the 3.20.0 release
To make sure the latest GEF release remains compatible with older
Eclipse release, we should not only build and test against the upcoming
release, but also against previous releases.

With this change, the BREE for all test bundles is reverted back to Java
17, in order to allow compilation and execution with a Java 17 JDK and
without having to configure a Maven toolchain.

Furthermore, the lower bound of all bundle dependencies is adjusted to
be compatible with the older releases.