Convert Zest tests into proper SWTBot tests
The current approach of executing the Zest-SWTBot tests is prone to
timing issues when interacting with the Draw2D Animation class.

To avoid this, the test suite itself should be executed in a separate
thread and all UI access should be done using the SWTBot API.

Note that SWTBot doesn't provide a Zest API, so a prototype defining the
required methods has been created as well.